+++ electron/preload.js
@@ -17,34 +17,4 @@ contextBridge.exposeInMainWorld("electronAPI", {
 
   // Open URL in default browser
   openExternal: (url) => ipcRenderer.invoke("open-external", url),
-
-  // Auto-updater methods
-  updater: {
-    // Check for updates
-    checkForUpdates: () => ipcRenderer.invoke("updater-check-for-updates"),
-
-    // Event listeners (one-way from main to renderer)
-    onChecking: (callback) => {
-      ipcRenderer.on("updater-checking", () => callback());
-      return () => ipcRenderer.removeAllListeners("updater-checking");
-    },
-
-    onUpdateAvailable: (callback) => {
-      ipcRenderer.on("updater-update-available", (_, info) => callback(info));
-      return () => ipcRenderer.removeAllListeners("updater-update-available");
-    },
-
-    onUpdateNotAvailable: (callback) => {
-      ipcRenderer.on("updater-update-not-available", (_, info) =>
-        callback(info)
-      );
-      return () =>
-        ipcRenderer.removeAllListeners("updater-update-not-available");
-    },
-
-    onError: (callback) => {
-      ipcRenderer.on("updater-error", (_, error) => callback(error));
-      return () => ipcRenderer.removeAllListeners("updater-error");
-    },
-  },
 });